+++ CHANGELOG.md
@@ -7,9 +7,7 @@
 
 ## v0.10.0 - dev
 
-- `QuantumClifffordJuMPExt`: New extension enabling:
-    - Compute minimum distance for quantum LDPC codes via Mixed Integer Programming (MIP) using `JuMP`.
-    - Introduce the `AbstractDistanceAlg` interface in `QECCore`.
+- `QuantumClifffordJuMPExt`: extension for computing the minimum distance for quantum LDPC codes via Mixed Integer Programming (MIP) using `JuMP`.
 - Improvements to `GeneralizedStabilizer` API:
     - Tensor product operations:
         - Between multiple `GeneralizedStabilizer` states: `genstab₁ ⊗ genstab₂ ⊗ ...`